requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
pypdf==3.17.4
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.5.2
huggingface-hub==0.20.1